Move save config back to train
