+++ chemprop/cli/common.py
@@ -1,6 +1,5 @@
 import logging
 from argparse import ArgumentParser, Namespace, ArgumentError
-from copy import deepcopy
 from pathlib import Path
 
 from chemprop.cli.utils import LookupAction
@@ -183,16 +182,3 @@ def process_common_args(args: Namespace) -> Namespace:
 def validate_common_args(args):
     pass
 
-
-def save_config(parser: ArgumentParser, args: Namespace, config_path: Path):
-    config_args = deepcopy(args)
-    for key, value in vars(config_args).items():
-        if isinstance(value, Path):
-            setattr(config_args, key, str(value))
-
-    for key in ["atom_features_path", "atom_descriptors_path", "bond_features_path"]:
-        if getattr(config_args, key) is not None:
-            for index, path in getattr(config_args, key).items():
-                getattr(config_args, key)[index] = str(path)
-
-    parser.write_config_file(parsed_namespace=config_args, output_file_paths=[str(config_path)])
